@@ -103,39 +103,40 @@
   },
   "private": false,
   "devDependencies": {
-    "@angular-devkit/build-angular": "16.2.0",
-    "@angular-devkit/core": "16.2.0",
-    "@angular-devkit/schematics": "16.2.0",
-    "@angular-eslint/eslint-plugin": "^16.2.0",
-    "@angular-eslint/eslint-plugin-template": "^16.2.0",
-    "@angular-eslint/template-parser": "^16.2.0",
-    "@angular/cli": "~16.2.0",
-    "@angular/compiler-cli": "16.2.1",
-    "@angular/language-service": "16.2.1",
-    "@nx/angular": "16.10.0",
-    "@nx/cypress": "16.10.0",
-    "@nx/eslint-plugin": "16.10.0",
-    "@nx/jest": "16.10.0",
-    "@nx/linter": "16.10.0",
-    "@nx/workspace": "16.10.0",
-    "@schematics/angular": "^16.2.6",
+    "@angular-devkit/build-angular": "18.1.2",
+    "@angular-devkit/core": "18.1.2",
+    "@angular-devkit/schematics": "18.1.2",
+    "@angular-eslint/eslint-plugin": "18.0.1",
+    "@angular-eslint/eslint-plugin-template": "18.0.1",
+    "@angular-eslint/template-parser": "18.0.1",
+    "@angular/cli": "~18.1.0",
+    "@angular/compiler-cli": "18.1.2",
+    "@angular/language-service": "18.1.2",
+    "@nx/angular": "19.5.3",
+    "@nx/cypress": "19.5.3",
+    "@nx/eslint": "19.5.3",
+    "@nx/eslint-plugin": "19.5.3",
+    "@nx/jest": "19.5.3",
+    "@nx/workspace": "19.5.3",
+    "@schematics/angular": "18.1.2",
     "@types/jest": "^29.5.6",
-    "@types/node": "^18.18.6",
-    "@typescript-eslint/eslint-plugin": "^6.8.0",
-    "@typescript-eslint/parser": "^6.8.0",
-    "cypress": "^13.0.0",
-    "eslint": "8.51",
-    "eslint-config-prettier": "^9.0.0",
+    "@types/node": "18.16.9",
+    "@typescript-eslint/eslint-plugin": "7.16.0",
+    "@typescript-eslint/parser": "7.16.0",
+    "@typescript-eslint/utils": "^7.16.0",
+    "autoprefixer": "^10.4.0",
+    "cypress": "13.13.0",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "9.0.0",
     "eslint-plugin-cypress": "2.15",
     "eslint-plugin-ngrx": "^2.1.3",
     "imng-kendo-schematics": "^6.233.2",
     "jest": "29.7",
     "jest-environment-jsdom": "29.7",
     "jest-junit": "^16.0.0",
-    "jest-preset-angular": "^13.1.2",
-    "ng-packagr": "^16.2.3",
-    "nx": "16.10.0",
-    "nx-cloud": "16.5.2",
+    "jest-preset-angular": "14.1.1",
+    "ng-packagr": "18.1.0",
+    "nx": "19.5.3",
     "openapi-mermaid": "^3.170.1",
     "openapi-ts-generator": "^6.170.3",
     "postcss": "^8.4.31",
@@ -145,26 +146,27 @@
     "prettier": "^3.0.3",
     "ts-jest": "^29.1.1",
     "ts-node": "10.9.1",
-    "typescript": "5.1.3"
+    "typescript": "5.5.4"
   },
   "dependencies": {
-    "@angular/animations": "16.2.1",
-    "@angular/common": "16.2.1",
-    "@angular/compiler": "16.2.1",
-    "@angular/core": "16.2.1",
-    "@angular/forms": "16.2.1",
-    "@angular/localize": "16.2.1",
-    "@angular/platform-browser": "16.2.1",
-    "@angular/platform-browser-dynamic": "16.2.1",
-    "@angular/router": "16.2.1",
-    "@angular/service-worker": "16.2.1",
+    "@angular/animations": "18.1.2",
+    "@angular/common": "18.1.2",
+    "@angular/compiler": "18.1.2",
+    "@angular/core": "18.1.2",
+    "@angular/forms": "18.1.2",
+    "@angular/localize": "18.1.2",
+    "@angular/platform-browser": "18.1.2",
+    "@angular/platform-browser-dynamic": "18.1.2",
+    "@angular/router": "18.1.2",
+    "@angular/service-worker": "18.1.2",
     "@ericblade/quagga2": "^1.7.0",
     "@microsoft/applicationinsights-web": "^3.0.3",
     "@microsoft/signalr": "^7.0.12",
-    "@ngrx/effects": "^16.0.1",
-    "@ngrx/router-store": "^16.0.1",
-    "@ngrx/store": "^16.0.1",
-    "@ngrx/store-devtools": "^16.0.1",
+    "@ngrx/effects": "18.0.0",
+    "@ngrx/operators": "^18.0.0",
+    "@ngrx/router-store": "18.0.0",
+    "@ngrx/store": "18.0.0",
+    "@ngrx/store-devtools": "18.0.0",
     "@progress/kendo-angular-buttons": "^14.0.1",
     "@progress/kendo-angular-charts": "^14.0.1",
     "@progress/kendo-angular-common": "^14.0.1",
