@@ -3,7 +3,11 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
-import { GridModule, ExcelModule, PDFModule } from '@progress/kendo-angular-grid';
+import {
+  GridModule,
+  ExcelModule,
+  PDFModule,
+} from '@progress/kendo-angular-grid';
 import { DialogModule } from '@progress/kendo-angular-dialog';
 import { DropDownsModule } from '@progress/kendo-angular-dropdowns';
 import { ImngKendoGridModule } from 'imng-kendo-grid';
@@ -16,11 +20,19 @@ import { customersFeature } from './+state/customer.reducer';
 import { CustomerEffects } from './+state/customer.effects';
 
 import { CustomerListComponent, CustomerListFacade } from './customers-list';
-import { CustomerAddComponent, CustomerEditComponent, CustomerApiService, CustomerCrudFacade } from './customers-crud';
-
+import {
+  CustomerAddComponent,
+  CustomerEditComponent,
+  CustomerApiService,
+  CustomerCrudFacade,
+} from './customers-crud';
 
 @NgModule({
-  declarations: [CustomerListComponent, CustomerAddComponent, CustomerEditComponent],
+  declarations: [
+    CustomerListComponent,
+    CustomerAddComponent,
+    CustomerEditComponent,
+  ],
   imports: [
     CommonModule,
     GridModule,
@@ -37,10 +49,6 @@ import { CustomerAddComponent, CustomerEditComponent, CustomerApiService, Custom
     StoreModule.forFeature(customersFeature),
     EffectsModule.forFeature([CustomerEffects]),
   ],
-  providers: [
-    CustomerListFacade,
-    CustomerCrudFacade,
-    CustomerApiService,
-  ],
+  providers: [CustomerListFacade, CustomerCrudFacade, CustomerApiService],
 })
-export class CustomersModule { }
+export class CustomersModule {}
