@@ -1,27 +1,38 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
-import { GridModule, ExcelModule, PDFModule } from '@progress/kendo-angular-grid';
-import { DialogModule } from '@progress/kendo-angular-dialog';
+import { StoreModule } from '@ngrx/store';
 import { DateInputsModule } from '@progress/kendo-angular-dateinputs';
+import { DialogModule } from '@progress/kendo-angular-dialog';
 import { DropDownsModule } from '@progress/kendo-angular-dropdowns';
-import { ImngKendoGridModule } from 'imng-kendo-grid';
-import { ImngKendoGridODataModule } from 'imng-kendo-grid-odata';
+import {
+  GridModule,
+  ExcelModule,
+  PDFModule,
+} from '@progress/kendo-angular-grid';
 import { ImngDataEntryDialogModule } from 'imng-kendo-data-entry';
+import { ImngKendoGridModule } from 'imng-kendo-grid';
 import { ImngKendoGridFilteringModule } from 'imng-kendo-grid-filtering';
+import { ImngKendoGridODataModule } from 'imng-kendo-grid-odata';
 
 import { ProductsRoutingModule } from './products.routing';
 import { productsFeature } from './+state/product.reducer';
 import { ProductEffects } from './+state/product.effects';
-
 import { ProductListComponent, ProductListFacade } from './products-list';
-import { ProductAddComponent, ProductEditComponent, ProductApiService, ProductCrudFacade  } from './products-crud';
-
+import {
+  ProductAddComponent,
+  ProductEditComponent,
+  ProductApiService,
+  ProductCrudFacade,
+} from './products-crud';
 
 @NgModule({
-  declarations: [ProductListComponent, ProductAddComponent, ProductEditComponent ],
+  declarations: [
+    ProductListComponent,
+    ProductAddComponent,
+    ProductEditComponent,
+  ],
   imports: [
     CommonModule,
     GridModule,
@@ -39,10 +50,6 @@ import { ProductAddComponent, ProductEditComponent, ProductApiService, ProductCr
     StoreModule.forFeature(productsFeature),
     EffectsModule.forFeature([ProductEffects]),
   ],
-  providers: [
-    ProductListFacade,
-    ProductCrudFacade,
-    ProductApiService, 
-  ],
+  providers: [ProductListFacade, ProductCrudFacade, ProductApiService],
 })
-export class ProductsModule { }
+export class ProductsModule {}
